add multiple image support for iterm2 and chafa
