+++ CMakeLists.txt
@@ -223,11 +223,9 @@ list(APPEND UEBERZUG_SOURCES
 
     "src/canvas.cpp"
     "src/canvas/chafa.cpp"
-    "src/canvas/sixel/sixel.cpp"
-    "src/canvas/sixel/stdout.cpp"
+    "src/canvas/sixel.cpp"
     "src/canvas/kitty/kitty.cpp"
     "src/canvas/kitty/chunk.cpp"
-    "src/canvas/kitty/window/stdout.cpp"
     "src/canvas/iterm2/iterm2.cpp"
     "src/canvas/iterm2/chunk.cpp"
