+++ src/canvas/chafa.cpp
@@ -29,7 +29,11 @@ void gstring_delete(GString* str) {
     g_string_free(str, true);
 }
 
-ChafaCanvas::ChafaCanvas()
+Chafa::Chafa(std::unique_ptr<Image> new_image, std::mutex& stdout_mutex):
+symbol_map(chafa_symbol_map_new()),
+config(chafa_canvas_config_new()),
+image(std::move(new_image)),
+stdout_mutex(stdout_mutex)
 {
     const auto envp = c_unique_ptr<gchar*, g_strfreev> {
         g_get_environ()
@@ -40,29 +44,13 @@ ChafaCanvas::ChafaCanvas()
     logger->warn("This canvas is meant to be used as a last resort. Please"
             " use the X11 output or switch to a terminal that has kitty,"
             " sixel or iterm2 support.");
-}
 
-ChafaCanvas::~ChafaCanvas()
-{
-    chafa_canvas_unref(canvas);
-    chafa_canvas_config_unref(config);
-    chafa_symbol_map_unref(symbol_map);
-    chafa_term_info_unref(term_info);
-}
-
-void ChafaCanvas::add_image(const std::string& identifier, std::unique_ptr<Image> new_image)
-{
-    remove_image(identifier);
-
-    image = std::move(new_image);
     const auto dims = image->dimensions();
     x = dims.x + 1;
     y = dims.y + 1;
     horizontal_cells = std::ceil(static_cast<double>(image->width()) / dims.terminal.font_width);
     vertical_cells = std::ceil(static_cast<double>(image->height()) / dims.terminal.font_height);
 
-    symbol_map = chafa_symbol_map_new();
-    config = chafa_canvas_config_new();
     chafa_symbol_map_add_by_tags(symbol_map, CHAFA_SYMBOL_TAG_BLOCK);
     chafa_symbol_map_add_by_tags(symbol_map, CHAFA_SYMBOL_TAG_BORDER);
     chafa_symbol_map_add_by_tags(symbol_map, CHAFA_SYMBOL_TAG_SPACE);
@@ -70,8 +58,22 @@ void ChafaCanvas::add_image(const std::string& identifier, std::unique_ptr<Image
     chafa_canvas_config_set_symbol_map(config, symbol_map);
     chafa_canvas_config_set_pixel_mode(config, CHAFA_PIXEL_MODE_SYMBOLS);
     chafa_canvas_config_set_geometry(config, horizontal_cells, vertical_cells);
-    canvas = chafa_canvas_new(config);
+}
 
+Chafa::~Chafa()
+{    
+    chafa_canvas_unref(canvas);
+    chafa_canvas_config_unref(config);
+    chafa_symbol_map_unref(symbol_map);
+    chafa_term_info_unref(term_info);
+
+    std::scoped_lock lock {stdout_mutex};
+    util::clear_terminal_area(x, y, horizontal_cells, vertical_cells);
+}
+
+void Chafa::draw()
+{
+    canvas = chafa_canvas_new(config);
     chafa_canvas_draw_all_pixels(canvas,
             CHAFA_PIXEL_BGRA8_UNASSOCIATED,
             image->data(),
@@ -79,11 +81,6 @@ void ChafaCanvas::add_image(const std::string& identifier, std::unique_ptr<Image
             image->height(),
             image->width() * 4);
 
-    draw();
-}
-
-void ChafaCanvas::draw()
-{
     const auto result = c_unique_ptr<GString, gstring_delete> {
 #if CHAFA_VERSION_CUR_STABLE >= 0x10e00
 #elif CHAFA_VERSION_CUR_STABLE >= 0x10c00
@@ -92,6 +89,8 @@ void ChafaCanvas::draw()
     };
     auto ycoord = y;
     const auto lines = util::str_split(result->str, "\n");
+
+    std::scoped_lock lock {stdout_mutex};
     util::save_cursor_position();
     std::ranges::for_each(std::as_const(lines), [&] (const auto& line) {
         util::move_cursor(ycoord++, x);
@@ -99,19 +98,4 @@ void ChafaCanvas::draw()
     });
     std::cout << std::flush;
     util::restore_cursor_position();
-}
-
-void ChafaCanvas::remove_image([[maybe_unused]] const std::string& identifier)
-{
-    if (horizontal_cells == 0 && vertical_cells == 0) {
-        return;
-    }
-
-    util::clear_terminal_area(x, y, horizontal_cells, vertical_cells);
-    chafa_canvas_unref(canvas);
-    chafa_canvas_config_unref(config);
-    chafa_symbol_map_unref(symbol_map);
-    canvas = nullptr;
-    config = nullptr;
-    symbol_map = nullptr;
 }
