+++ src/canvas/iterm2/iterm2.cpp
@@ -18,6 +18,7 @@
 #include "chunk.hpp"
 #include "util.hpp"
 #include "dimensions.hpp"
+#include "terminal.hpp"
 #include "image.hpp"
 
 #include <fstream>
@@ -33,25 +34,27 @@
 
 namespace fs = std::filesystem;
 
-Iterm2Canvas::Iterm2Canvas()
+Iterm2::Iterm2(std::unique_ptr<Image> new_image, std::mutex& stdout_mutex):
+image(std::move(new_image)),
+stdout_mutex(stdout_mutex)
 {
     logger = spdlog::get("iterm2");
     logger->info("Canvas created");
-}
 
-void Iterm2Canvas::add_image(const std::string& identifier, std::unique_ptr<Image> new_image)
-{
-    remove_image(identifier);
-    image = std::move(new_image);
     const auto dims = image->dimensions();
     x = dims.x + 1;
     y = dims.y + 1;
-    max_width = dims.max_w;
-    max_height = dims.max_h;
-    draw();
+    horizontal_cells = std::ceil(static_cast<double>(image->width()) / dims.terminal.font_width);
+    vertical_cells = std::ceil(static_cast<double>(image->height()) / dims.terminal.font_height);
 }
 
-void Iterm2Canvas::draw()
+Iterm2::~Iterm2()
+{
+    std::scoped_lock lock {stdout_mutex};
+    util::clear_terminal_area(x, y, horizontal_cells, vertical_cells);
+}
+
+void Iterm2::draw()
 {
     str.append("\033]1337;File=inline=1;");
     const auto filename = image->filename();
@@ -79,13 +82,7 @@ void Iterm2Canvas::draw()
     str.clear();
 }
 
-void Iterm2Canvas::remove_image([[maybe_unused]] const std::string& identifier)
-{
-    util::clear_terminal_area(x, y, max_width, max_height);
-    image.reset();
-}
-
-auto Iterm2Canvas::process_chunks(const std::string_view filename, int chunk_size, size_t num_bytes) -> std::vector<std::unique_ptr<Iterm2Chunk>>
+auto Iterm2::process_chunks(const std::string_view filename, int chunk_size, size_t num_bytes) -> std::vector<std::unique_ptr<Iterm2Chunk>>
 {
     const int num_chunks = std::ceil(static_cast<double>(num_bytes) / chunk_size);
     std::vector<std::unique_ptr<Iterm2Chunk>> chunks;
